test(reborn): address review round 2 (de-mask failure coverage, register ordering, dead field, doc + file-size)
Multi-agent review findings on #5465:

- T1 (de-mask failure coverage): add group scenario
  scenario_failure_category_demasked -- under-scripts a thread so the
  scripted model exhausts -> ModelError -> Failed, and asserts the run's
  failure category is the TRUE "model_error", not the masked
  "driver_protocol_violation". Mutation-verified: removing
  with_checkpoint_state_store flips the category back to the sentinel.
  This proves the group-level de-mask wiring, which only loop_exit_applier
  unit tests covered before.
- M1 (register ordering): RebornThreadBuilder::build now registers the
  scoped gateway AFTER all fallible setup, so a failed build() can't leave
  a stale registration that trips the duplicate-registration panic on retry.
- M2 (dead field): drop GroupSharedStorage::checkpoint_state_store -- it was
  stored but never read; the live owners (evidence port + planned runtime)
  wire from the local var in into_group.
- C1 (file size): relocate the build_compaction_ports test out of
  loop_driver_host.rs (3504 -> 3344) into a sibling
  loop_driver_host/compaction_tests.rs via #[cfg(test)] #[path], matching
  the existing port_adapter_tests pattern (private access preserved). Keeps
  the prod diff under the 200-line justification trigger; pre-existing
  >3000-line condition is tracked in #4088.
- LP1/LP2: fix stale docs -- main.rs scenario count (three -> five) and the
  thread() method doc ("runtime builder" -> per-thread workflow builder).

Verified: all reborn_group_* green, loop_driver_host lib tests green
(incl. relocated compaction_tests), clippy -D warnings clean, fmt clean.
